Update the heap (#158)
This will fix issue #150 and allow us to close off #149 without merging.

The impact here is a more deterministic processing model, where a document can only make it into the top-k heap if its current score exceeds the heap threshold. This means that, given a query in which all document scores are tied, we would return the first k matching documents.