Paralelize algorithms (#473)
* Remove PSTL

At this stage, all algorithms that used PSTL are turned into sequential.
Further effort is needed to implement parallelization directly via TBB.

* Add Docker-based unit test CI workflow

* Fix container name

* Add GCC 10, 9, and 8 builds to CI

* Add clang11 build to CI

* Fix syntax

* Fix Gumbo compilation

* Add more CI tests

* Update Catch2 to v2.13.7

* Implement conditional parallel transform.

* Fix header compilation

* CI: Compile headers in docker

* Change condition for including parallel algorithms

* Parallelize algorithms

Uses algorithms in pisa namespace that accept an execution, and either
pass it to the c++17 version of the algorithm if implemented in the
standard library, or ignores it otherwise. These algorithms won't be
paralelized if the std lib used doesn't support them.