Whitespace tokenizer
Implements a whitespace tokenizer next to the old term tokenizer. The
old tokenizer is renamed to EnglishTokenizer (as it contains
English-specific rules such as possessives), and both tokenizers are now
organized into a class hierarchy with a common virtual interface, so
that they could be used interchangeably in the future, as well so that
perhaps more tokenizers can be implemented.