Custom sift down for binary heap (#507)
* Add google benchmark submodule

* Custom sift down for binary heap

The previous solution where we swap first and last elements of the heap
and call `pop_heap` is incorrect according to the standard, and fails
when compiled with libc++ 15. Calling `push_heap` in addition to
`pop_heap` is less efficient; in fact, the custom sift down solution is
more efficient than the previous implementation (according to the
microbenchmarks).