Remove structured binding in favor of two initializations
It's a known issue that when using a structured binding, we cannot pass
those variables to a lambda. This is why the lambda capture below worked
around it with [var = var] hack. However, this is cluttered and
confusing, so I removed the use of structured binding so that we can
pass those values to the lambda just by name.