Remove use of nullptr static cast
We used some shady cast that involved casting a nullptr, in order to
pass lambdas to an object and keep it copy-assignable/copyable. Instead,
I now store the function objects in the parent object. This means that
if we simply pass a lambda, the struct won't be copyable. The solution
for that is to append + in front of the lambda, which casts it into a
function pointer.