Query refactoring
Weights are now stored together with term IDs and resolved at
construction time according to one of the policies. In our tools, we use
the default policy that removes duplicates and sets the weight to the
number of occurrences of the term in a query. Other policies are, for
the time being, only available programmatically via the library API.

Some legacy code used to parse and process queries has been removed in
favor of the text analyzer and the new query parser.

Because weights are resolved when a query object is created, I also
refactored creating the cursors: now the weight is simply taken from the
query.