Rework `compute_intersection` query length filter
This moves the query length filter out of the main
`compute_intersection` binary for automated testing.

- Enable `tools/tests` in cmake build
- Fix old tests in `tools/tests/test_app.cpp`
- Add `args::QueryFilter` tests for `compute_intersection`

The `tools/tests` are not yet configured for CI builds as they require
both options `PISA_BUILD_TOOLS=ON` and `PISA_ENABLE_TESTING=ON`.
Although, they can be run locally. This will be resolved in a later PR
(see #620).

See #392